Brute force program with an action menu
Hello! after hours of me staring at my screen as the code fails, i present to you my first ever program that ive personally coded and NOT stole. 
yes i did use AI but im not trying to spend 30 minutes to fix a syntax error so here you go
I would love for you guys to edit it and make it much better but give me credits i really worked hard on this

this is version 1.0 if i ever think of updating this mess but sure 
im new to batch and so far i  love it so give me tips and hints that i can improve on this 
anyways enjoy!

Shout out to Ebola man yt for inspiring me to do a similar thing here but hey i added an interactive menu
so im original C: